Changed the way our stack works to split it and follow a more logical structure. Removed unused packages as well.  @tanstack/zod-form-adapter and  @tanstack/react-form
