+++ __tests__/DataProvider.test.tsx
@@ -1,6 +1,6 @@
-import { renderHook, act, render, screen } from "@testing-library/react-native";
+import { act, render, renderHook, screen } from "@testing-library/react-native";
 import DateProvider, { useDate } from "../providers/DateProvider";
-import { View, Text } from "react-native";
+import { Text, View } from "react-native";
 
 describe("DateProvider and useDate", () => {
   it("should update selectedDate when setSelectedDate is called", () => {
