+++ __tests__/useWeek.test.ts
@@ -1,4 +1,4 @@
-import { renderHook, act } from "@testing-library/react-native";
+import { act, renderHook } from "@testing-library/react-native";
 import useWeek from "../hooks/useWeek";
 
 describe("useWeek hook", () => {
