+++ app/auth/profile/editprofile.tsx
@@ -1,39 +1,16 @@
-import {
-    StyleSheet,
-    Text,
-    TextInput,
-    TouchableOpacity,
-    View,
-  } from "react-native";
-  import FieldInfo from "../../../components/FieldInfo";
-  import React from "react";
-  import { z } from "zod";
-  import { useForm } from "@tanstack/react-form";
-  import { zodValidator } from "@tanstack/zod-form-adapter";
-  import useProfile from "../../../hooks/useProfile";
-  import { useToast } from "../../../providers/ToastProvider";
-  import {
-    colors,
-    ScaleSize,
-    ScaleSizeH,
-    ScaleSizeW,
-  } from "../../../utils/SharedStyles";
-import { router } from "expo-router";
-  
-  const schema = z.object({
-import { zodResolver } from "@hookform/resolvers/zod";
-import { router } from "expo-router";
-import React from "react";
-import { useForm } from "react-hook-form";
 import { StyleSheet, Text, TouchableOpacity } from "react-native";
+import React from "react";
 import { z } from "zod";
-import SubmitButton from "../components/Forms/SubmitButton";
-import FormField from "../components/Forms/TextInput";
-import useProfile from "../hooks/useProfile";
-import { useToast } from "../providers/ToastProvider";
-import { colors, ScaleSize, ScaleSizeH, ScaleSizeW } from "../utils/SharedStyles";
-import FormContainer from "../components/Forms/FormContainer";
-import FormHeader from "../components/Forms/FormHeader";
+import useProfile from "../../../hooks/useProfile";
+import { useToast } from "../../../providers/ToastProvider";
+import { colors, ScaleSize, ScaleSizeH, ScaleSizeW } from "../../../utils/SharedStyles";
+import { router } from "expo-router";
+import { useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import FormField from "../../../components/Forms/TextInput";
+import FormContainer from "../../../components/Forms/FormContainer";
+import FormHeader from "../../../components/Forms/FormHeader";
+import SubmitButton from "../../../components/Forms/SubmitButton";
 
 const schema = z
   .object({
