+++ app/auth/profile/organisation/editactivity.tsx
@@ -1,10 +1,9 @@
 import { useLocalSearchParams } from "expo-router";
 import ActivityEdit from "../../../../components/weekoverview_components/activity_components/ActivityEdit";
-import ActivityEdit from "../components/weekoverview_components/activity_components/ActivityEdit";
 import { useQueryClient } from "@tanstack/react-query";
-import { useDate } from "../providers/DateProvider";
-import { dateToQueryKey } from "../hooks/useActivity";
-import { ActivityDTO } from "../DTO/activityDTO";
+import { useDate } from "../../../../providers/DateProvider";
+import { dateToQueryKey } from "../../../../hooks/useActivity";
+import { ActivityDTO } from "../../../../DTO/activityDTO";
 
 type Params = {
   activityId: string;
