+++ app/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import SplashScreen from "../components/SplashScreen";
 import { router } from "expo-router";
