+++ components/weekoverview_components/DaysContainer.tsx
@@ -1,4 +1,4 @@
-import { View, StyleSheet } from "react-native";
+import { StyleSheet, View } from "react-native";
 import { GestureDetector } from "react-native-gesture-handler";
 import Animated from "react-native-reanimated";
 import { DAYS_OF_WEEK } from "../../constants/daysOfWeek";
