+++ components/weekoverview_components/activity_components/ActivityItemHeader.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { View, Text, StyleSheet } from "react-native";
+import { StyleSheet, Text, View } from "react-native";
 import { colors, ScaleSize, ScaleSizeH } from "../../../utils/SharedStyles";
 
 /**
