+++ package.json
@@ -35,9 +35,7 @@
     "@react-native-async-storage/async-storage": "1.23.1",
     "@react-native-community/datetimepicker": "8.0.1",
     "@react-native-picker/picker": "2.7.5",
-    "@tanstack/react-form": "^0.34.4",
     "@tanstack/react-query": "^5.56.2",
-    "@tanstack/zod-form-adapter": "^0.34.4",
     "@types/react-native-base64": "^0.2.2",
     "expo": "~51.0.38",
     "expo-constants": "~16.0.2",
